refactor: local 패키지 생성, local.model 패키지 생성
