+++ app/src/main/java/com/sdhong/pokemonapp/local/Pokemons.kt
@@ -1,0 +1,8 @@
+package com.sdhong.pokemonapp.local
+
+import com.sdhong.pokemonapp.local.model.Pokemon
+
+object Pokemons {
+
+    val historyPokemons = mutableListOf<Pokemon.History>()
+}
