--- app/src/main/java/com/sdhong/pokemonapp/model/Pokemons.kt
@@ -1,6 +1,0 @@
-package com.sdhong.pokemonapp.model
-
-object Pokemons {
-
-    val historyPokemons = mutableListOf<Pokemon.History>()
-}
