+++ app/src/main/java/com/sdhong/pokemonapp/remote/api/PokemonApi.kt
@@ -4,7 +4,6 @@ import com.sdhong.pokemonapp.remote.model.PokemonDetailResponse
 import com.sdhong.pokemonapp.remote.model.PokemonListResponse
 import retrofit2.http.GET
 import retrofit2.http.Path
-import retrofit2.http.Query
 
 interface PokemonApi {
